+++ core/src/hasher.rs
@@ -1,3 +1,5 @@
+//! Custom hasher implementations
+
 /// The hasher used to compare layouts.
 #[derive(Debug)]
 pub struct Hasher(twox_hash::XxHash64);
