+++ core/src/lib.rs
@@ -22,6 +22,8 @@ pub mod time;
 mod background;
 mod color;
 mod font;
+mod hasher;
+pub mod image;
 mod length;
 mod padding;
 mod point;
@@ -33,6 +35,7 @@ pub use alignment::Alignment;
 pub use background::Background;
 pub use color::Color;
 pub use font::Font;
+pub use hasher::Hasher;
 pub use length::Length;
 pub use padding::Padding;
 pub use point::Point;
